feat: Force MaxAI to navigate to new pages more often
Max sometimes just doesn't do anything when asked to create surveys, feature flags, experiments, etc. because it doesn't have access to that page's tools from any page. Max *could* just navigate to that page and then get access to an extra set of tools but it doesn't usually do that.

Let's tweak our prompt slightly to suggest to Max it should move to a different page to get access to extra set of tools